Make 'onProgress' part of the public API
